Rename to use homewizard-v2 prefix for templates
