+++ meter/homewizard-battery.go
@@ -48,7 +48,7 @@ func NewHomeWizardBatteryFromConfig(other map[string]any) (api.Meter, error) {
 
 	// Validate required parameters
 	if cc.Host == "" || cc.Token == "" {
-		return nil, fmt.Errorf("missing host or token - run 'evcc token homewizard-v2'")
+		return nil, fmt.Errorf("missing host or token - run 'evcc token homewizard'")
 	}
 
 	if cc.Controller == "" {
