+++ templates/definition/meter/homewizard-kwh.yaml
@@ -2,40 +2,12 @@ template: homewizard-kwh
 products:
   - brand: HomeWizard
     description:
-      generic: kWh Meter (HWE-KWH1/3)
-      de: kWh-Zähler (HWE-KWH1/3)
-requirements:
-  description:
-    en: |
-      kWh Meter for PV monitoring using WebSocket API v2.
-
-      **Setup:** Run `evcc token homewizard-v2` to discover and pair devices
-    de: |
-      kWh-Zähler für PV-Messung mit WebSocket API v2.
-
-      **Einrichtung:** Führen Sie `evcc token homewizard-v2` aus
+      generic: kWh Meter
 params:
+  - name: usage
+    choice: ["pv", "charge"]
   - name: host
-    description:
-      en: Device hostname or IP
-      de: Gerätehostname oder IP
-    required: true
-  - name: token
-    description:
-      en: Device authentication token
-      de: Geräte-Authentifizierungstoken
-    required: true
-    mask: true
-  - name: timeout
-    description:
-      en: Connection timeout
-      de: Verbindungs-Timeout
-    advanced: true
-    default: 30s
 render: |
-  type: homewizard-kwh
-  host: {{ .host }}
-  token: {{ .token }}
-  {{- if .timeout }}
-  timeout: {{ .timeout }}
-  {{- end }}
+  type: homewizard
+  uri: http://{{ .host }}
+  usage: {{ .usage }}
