+++ templates/definition/meter/homewizard-v2-battery.yaml
@@ -1,4 +1,4 @@
-template: homewizard-battery
+template: homewizard-v2-battery
 products:
   - brand: HomeWizard
     description:
@@ -9,14 +9,14 @@ requirements:
     en: |
       Battery device for SoC and power monitoring using WebSocket API v2.
 
-      **Setup:** Run `evcc token homewizard-v2` to discover and pair devices
-      
+      **Setup:** Run `evcc token homewizard` to discover and pair devices
+
       **Note:** Requires a `homewizard-p1` meter (grid) for battery control
     de: |
       Batteriegerät für SoC und Leistungsmessung mit WebSocket API v2.
 
-      **Einrichtung:** Führen Sie `evcc token homewizard-v2` aus
-      
+      **Einrichtung:** Führen Sie `evcc token homewizard` aus
+
       **Hinweis:** Benötigt einen `homewizard-p1` Zähler (Netz) für Batteriesteuerung
 params:
   - name: host
